Update PerfectNumber.java
Code is incomplete because it Need to Check for Prefect Number. Its wasn't checking for weather the count is equal to sum of all divisor.
For Example number=10, Divisor are 1 , 2 ,5 which equals to 8. which is not equal to number. That means It is not a prefect number. it should than return false .